* 'master' of https://gitee.com/openarkcompiler/OpenArkCompiler:
  fixed usage of  abs()
  Fix more type compatibility issues
  Fix bug introduced by last commit
  changed BB::frequency field from uint32 to uint64
  Fix more type compatibility issues
  Fix bug introduced by last commit
  changed BB::frequency field from uint32 to uint64
  Fix more type compatibility issues
  Fix bug introduced by last commit
  changed BB::frequency field from uint32 to uint64

+++ src/mapleall/maple_me/src/hdse.cpp
@@ -250,12 +250,11 @@ void HDSE::RemoveNotRequiredStmtsInBB(BB &bb) {
         }
         bb.SetKind(kBBFallthru);
         if (UpdateFreq()) {
-          int64_t succ0Freq = static_cast<int64_t>(bb.GetSuccFreq()[0]);
+          uint64 succ0Freq = static_cast<int64_t>(bb.GetSuccFreq()[0]);
           bb.GetSuccFreq().resize(1);
           bb.SetSuccFreq(0, bb.GetFrequency());
           ASSERT(bb.GetFrequency() >= succ0Freq, "sanity check");
-          bb.GetSucc(0)->SetFrequency(static_cast<uint32>(bb.GetSucc(0)->GetFrequency() +
-            (bb.GetFrequency() - succ0Freq)));
+          bb.GetSucc(0)->SetFrequency(bb.GetSucc(0)->GetFrequency() + (bb.GetFrequency() - succ0Freq));
         }
       }
       // A ivar contained in stmt
@@ -281,7 +280,7 @@ void HDSE::RemoveNotRequiredStmtsInBB(BB &bb) {
       bool isPme = mirModule.CurFunction()->GetMeFunc()->GetPreMeFunc() != nullptr;
       if (mestmt->IsCondBr() && !isPme) {  // see if foldable to unconditional branch
         CondGotoMeStmt *condbr = static_cast<CondGotoMeStmt*>(mestmt);
-        int64_t removedFreq = 0;
+        uint64 removedFreq = 0;
         if (!mirModule.IsJavaModule() && condbr->GetOpnd()->GetMeOp() == kMeOpConst) {
           CHECK_FATAL(IsPrimitiveInteger(condbr->GetOpnd()->GetPrimType()),
                       "MeHDSE::DseProcess: branch condition must be integer type");
@@ -317,7 +316,7 @@ void HDSE::RemoveNotRequiredStmtsInBB(BB &bb) {
           if (UpdateFreq()) {
             bb.GetSuccFreq().resize(1);
             bb.SetSuccFreq(0, bb.GetFrequency());
-            bb.GetSucc(0)->SetFrequency(static_cast<uint32>(bb.GetSucc(0)->GetFrequency() + removedFreq));
+            bb.GetSucc(0)->SetFrequency(bb.GetSucc(0)->GetFrequency() + removedFreq);
           }
         } else {
           DetermineUseCounts(condbr->GetOpnd());
